✨ feat(tests): refactor SeededUserAccountsEmails type definition for clarity
